+++ phonegap-build/platforms/android/assets/www/src/app/dashboard/chart.js
@@ -18,7 +18,8 @@ angular.module('HABmin.chart', [
     'HABmin.chartSave',
     'HABmin.iconModel',
     'ngVis',
-    'ngConfirmClick'
+    'ngConfirmClick',
+    'ResizePanel'
 ])
 
     .config(function config($stateProvider) {
@@ -41,7 +42,7 @@ angular.module('HABmin.chart', [
     })
 
     .controller('DashboardChartCtrl',
-    function DashboardChartCtrl($scope, $modal, locale, PersistenceItemModel, PersistenceServiceModel, PersistenceDataModel, ChartListModel, ChartSave, growl, VisDataSet, $interval, $timeout) {
+    function DashboardChartCtrl($scope, locale, PersistenceItemModel, PersistenceServiceModel, PersistenceDataModel, ChartListModel, ChartSave, growl, VisDataSet, $interval, $timeout) {
         var itemsLoaded = 0;
         var itemsLoading = 0;
         var newChart;
@@ -805,44 +806,6 @@ angular.module('HABmin.chart', [
 
             return curData;
         }
-    })
-
-    .directive('resizePage', function ($window) {
-        return function ($scope, element) {
-            var w = angular.element($window);
-            $scope.getWindowDimensions = function () {
-                return {
-                    'h': w.height()
-                };
-            };
-            $scope.$watch($scope.getWindowDimensions, function (newValue, oldValue) {
-                $scope.windowHeight = newValue.h;
-                $scope.styleItemList = function () {
-                    return {
-                        'height': (newValue.h - 232) + 'px'
-                    };
-                };
-                $scope.styleChartList = function () {
-                    return {
-                        'height': (newValue.h - 165) + 'px'
-                    };
-                };
-                $scope.styleChartPanel = function () {
-                    return {
-                        'height': (newValue.h - 93) + 'px'
-                    };
-                };
-                $scope.styleChart = function () {
-                    return {
-                        'height': (newValue.h - 132) + 'px'
-                    };
-                };
-            }, true);
-
-            w.bind('resize', function () {
-                $scope.$apply();
-            });
-        };
     })
 
 ;