+++ phonegap-build/platforms/android/assets/www/src/app/models/itemModel.js
@@ -8,10 +8,11 @@
  * (c) 2014 Chris Jackson (chris@cd-jackson.com)
  */
 angular.module('HABmin.itemModel', [
+    'HABmin.userModel'
 ])
 
-    .service('ItemModel', function ($http, $q) {
-        this.url = HABminServer + '/rest/items';
+    .service('ItemModel', function ($http, $q, UserService) {
+        this.url = UserService.getServer() + '/rest/items';
 
         this.socket = null;
         this.sendCommand = function (item, value) {