+++ phonegap-build/platforms/android/assets/www/src/app/user/userGeneral.js
@@ -8,13 +8,20 @@
  * (c) 2014 Chris Jackson (chris@cd-jackson.com)
  */
 angular.module('UserGeneralPrefs', [
-    'ngLocalize'
+    'ngLocalize',
+    'HABmin.userModel'
 ])
     .service('UserGeneralPrefs',
-    function ($modal, $rootScope) {
+    function ($modal, $rootScope, UserService) {
         this.showModal = function () {
+            var scope = $rootScope.$new();
+            scope.model = {};
+            scope.model.theme = UserService.getTheme();
+
             var controller = function ($scope, $modalInstance) {
                 $scope.ok = function (result) {
+                    UserService.setTheme(scope.model.theme);
+
                     $modalInstance.close(result);
                 };
                 $scope.cancel = function (result) {
@@ -22,15 +29,13 @@ angular.module('UserGeneralPrefs', [
                 };
             };
 
-            var scope = $rootScope.$new();
-            scope.embedded = window.PhoneGap;
-
             return $modal.open({
                 backdrop: 'static',
                 keyboard: true,
                 modalFade: true,
                 templateUrl: 'user/userGeneral.tpl.html',
                 controller: controller,
+                windowClass: UserService.getTheme(),
                 scope: scope
             }).result;
         };