+++ phonegap-build/www/templates-app.js
@@ -2,6 +2,7 @@ angular.module('templates-app', ['automation/rule.tpl.html', 'automation/schedul
 
 angular.module("automation/rule.tpl.html", []).run(["$templateCache", function($templateCache) {
   $templateCache.put("automation/rule.tpl.html",
+    "<!DOCTYPE html>\n" +
     "<div class=\"col-sm-4\">\n" +
     "    <div class=\"panel panel-default\">\n" +
     "        <!-- Header -->\n" +
@@ -23,7 +24,7 @@ angular.module("automation/rule.tpl.html", []).run(["$templateCache", function($
     "        </div>\n" +
     "\n" +
     "        <!-- Body -->\n" +
-    "        <div class=\"list-group\" ng-style=\"styleRuleList()\" ng-if=\"!selectCharts\" resize-page>\n" +
+    "        <div class=\"list-group\" ng-if=\"!selectCharts\" resize-panel>\n" +
     "            <div ng-repeat=\"choice in rules\">\n" +
     "                <a role=\"presentation\" ng-class=\"{'list-group-item':true, 'active':choice.id==selectedRule.id}\"\n" +
     "                   ng-confirm-click=\"selectRule(choice)\" ng-confirm-if=\"$parent.isDirty\"\n" +
@@ -43,13 +44,14 @@ angular.module("automation/rule.tpl.html", []).run(["$templateCache", function($
     "                <span i18n=\"habmin.ruleNoRules\"></span>\n" +
     "            </div>\n" +
     "        </div>\n" +
-    "        <div class=\"panel-footer panel-footer-small\">\n" +
+    "        <div class=\"panel-footer panel-footer-none\">\n" +
     "        </div>\n" +
     "    </div>\n" +
     "</div>\n" +
     "\n" +
+    "<!-- Rule Pane -->\n" +
     "<div class=\"col-sm-8\">\n" +
-    "    <div class=\"panel panel-default\" ng-style=\"styleEditor()\" resize-page>\n" +
+    "    <div class=\"panel panel-default\">\n" +
     "        <div class=\"panel-heading\" style=\"padding:3px;\">\n" +
     "            <button type=\"button\" ng-class=\"{'btn btn-sm btn-default':true, 'disabled':!isDirty}\" ng-click=\"saveRule()\">\n" +
     "                <span class=\"fa fa-floppy-o\"></span>\n" +
@@ -62,28 +64,23 @@ angular.module("automation/rule.tpl.html", []).run(["$templateCache", function($
     "            </button>\n" +
     "\n" +
     "            <div class=\"pull-right\">\n" +
-    "                <button ng-class=\"{'btn btn-sm btn-default':true, 'active':editSource==false}\" ng-click=\"showRule()\">\n" +
+    "                <button type=\"button\" ng-class=\"{'btn btn-sm btn-default':true, 'active':editSource==false}\" ng-click=\"showRule()\">\n" +
     "                    <span class=\"fa fa-puzzle-piece\"></span>\n" +
     "                    <span i18n=\"habmin.ruleRule\"></span>\n" +
     "                </button>\n" +
-    "                <button ng-class=\"{'btn btn-sm btn-default':true, 'active':editSource==true}\" ng-click=\"showSource()\">\n" +
+    "                <button type=\"button\" ng-class=\"{'btn btn-sm btn-default':true, 'active':editSource==true}\" ng-click=\"showSource()\">\n" +
     "                    <span class=\"fa fa-code\"></span>\n" +
     "                    <span i18n=\"habmin.ruleSource\"></span>\n" +
     "                </button>\n" +
     "            </div>\n" +
     "        </div>\n" +
     "\n" +
-    "        <ng-blockly ng-show=\"!editSource\">\n" +
-    "        </ng-blockly>\n" +
+    "        <div resize-panel>\n" +
+    "            <ng-blockly ng-show=\"!editSource\">\n" +
+    "            </ng-blockly>\n" +
     "\n" +
-    "        <div ng-show=\"editSource\" ng-model=\"codeEditor\" ui-ace=\"{\n" +
-    "              useWrapMode : true,\n" +
-    "              showGutter: true,\n" +
-    "              theme:'tomorrow_night_bright',\n" +
-    "              mode: 'openhabrules',\n" +
-    "              firstLineNumber: 5\n" +
-    "             }\"\n" +
-    "             ng-style=\"styleEditor()\">\n" +
+    "            <div style=\"height:100%\" ng-show=\"editSource\" ng-model=\"codeEditor\" ui-ace=\"aceOptions\">\n" +
+    "            </div>\n" +
     "        </div>\n" +
     "    </div>\n" +
     "</div>\n" +
@@ -161,6 +158,7 @@ angular.module("automation/scheduler.tpl.html", []).run(["$templateCache", funct
 
 angular.module("binding/zwave.tpl.html", []).run(["$templateCache", function($templateCache) {
   $templateCache.put("binding/zwave.tpl.html",
+    "<!DOCTYPE html>\n" +
     "<!-- Device List -->\n" +
     "<div class=\"col-sm-4\">\n" +
     "    <div class=\"panel panel-default\">\n" +
@@ -204,7 +202,7 @@ angular.module("binding/zwave.tpl.html", []).run(["$templateCache", function($te
     "        </div>\n" +
     "\n" +
     "        <!-- Body -->\n" +
-    "        <div class=\"list-group habmin-list\" ng-style=\"styleList()\" resize-page1>\n" +
+    "        <div class=\"list-group habmin-list\" resize-panel>\n" +
     "            <div ng-repeat=\"choice in devices\">\n" +
     "                <a role=\"presentation\" ng-class=\"{'list-group-item':true, 'active': choice.device==devEdit.device}\"\n" +
     "                   ng-click=\"selectDevice(choice)\">\n" +
@@ -229,7 +227,7 @@ angular.module("binding/zwave.tpl.html", []).run(["$templateCache", function($te
     "                    <span class=\"pull-right\">\n" +
     "                        <span popover=\"{{choice.powerInfo}}\" popover-trigger=\"mouseenter\" popover-placement=\"top\"\n" +
     "                              popover-append-to-body=\"true\" popover-popup-delay=\"500\">\n" +
-    "                            <span ng-class=\"{'text-success': choice.batteryLevel>='40','text-danger': choice.batteryLevel=='0','text-warning': choice.batteryLevel =='20','text-muted': choice.batteryLevel =='-1'}\">\n" +
+    "                            <span ng-class=\"{'text-success': choice.batteryLevel>=40,'text-danger': choice.batteryLevel==0,'text-muted': choice.batteryLevel=='UNK','text-warning': choice.batteryLevel<40}\">\n" +
     "                                <span ng-class=\"choice.batteryIcon\"></span>\n" +
     "                            </span>\n" +
     "                        </span>\n" +
@@ -246,11 +244,10 @@ angular.module("binding/zwave.tpl.html", []).run(["$templateCache", function($te
     "                <span class=\"fa fa-exclamation-triangle\"></span>\n" +
     "                <span i18n=\"zwave.zwaveNoDevices\"></span>\n" +
     "            </div>\n" +
-    "\n" +
     "        </div>\n" +
     "\n" +
     "        <!-- Footer -->\n" +
-    "        <div class=\"panel-footer panel-footer-small\">\n" +
+    "        <div class=\"panel-footer panel-footer-none\">\n" +
     "        </div>\n" +
     "    </div>\n" +
     "</div>\n" +
@@ -304,15 +301,16 @@ angular.module("binding/zwave.tpl.html", []).run(["$templateCache", function($te
     "                        </a>\n" +
     "                    </li>\n" +
     "                </ul>\n" +
-    "                <button ng-show=\"deviceDisplay=='CONFIG'\" type=\"button\" ng-click=\"deviceSave()\" ng-class=\"{'btn btn-default':true, 'disabled':!isDirty}\">\n" +
-    "                    <span class=\"fa fa-floppy-o\"></span>\n" +
-    "                    <span class=\"hidden-xs\" i18n=\"common.save\"></span>\n" +
-    "                </button>\n" +
-    "                <button ng-show=\"deviceDisplay=='CONFIG'\" type=\"button\" ng-click=\"deviceCancel()\" ng-class=\"{'btn btn-default':true, 'disabled':!isDirty}\">\n" +
-    "                    <span class=\"fa fa-times\"></span>\n" +
-    "                    <span class=\"hidden-xs\" i18n=\"common.cancel\"></span>\n" +
-    "                </button>\n" +
     "            </div>\n" +
+    "\n" +
+    "            <button ng-show=\"deviceDisplay=='CONFIG'\" type=\"button\" ng-click=\"deviceSave()\" ng-class=\"{'btn btn-default':true, 'disabled':!isDirty}\">\n" +
+    "                <span class=\"fa fa-floppy-o\"></span>\n" +
+    "                <span class=\"hidden-xs\" i18n=\"common.save\"></span>\n" +
+    "            </button>\n" +
+    "            <button ng-show=\"deviceDisplay=='CONFIG'\" type=\"button\" ng-click=\"deviceCancel()\" ng-class=\"{'btn btn-default':true, 'disabled':!isDirty}\">\n" +
+    "                <span class=\"fa fa-times\"></span>\n" +
+    "                <span class=\"hidden-xs\" i18n=\"common.cancel\"></span>\n" +
+    "            </button>\n" +
     "\n" +
     "            <div ng-hide=\"!devEdit.label\" class=\"btn-group pull-right\">\n" +
     "                <div class=\"text-right\">{{devEdit.label}}</div>\n" +
@@ -321,7 +319,7 @@ angular.module("binding/zwave.tpl.html", []).run(["$templateCache", function($te
     "        </div>\n" +
     "\n" +
     "        <!-- Body -->\n" +
-    "        <div ng-show=\"deviceDisplay=='CONFIG'\" class=\"list-group habmin-list\" ng-style=\"styleList()\">\n" +
+    "        <div ng-show=\"deviceDisplay=='CONFIG'\" class=\"list-group habmin-list\" resize-panel>\n" +
     "            <!-- INFORMATION -->\n" +
     "            <div ng-show=\"devEdit.information\">\n" +
     "                <a role=\"presentation\" class=\"list-group-item\" ng-click=\"showPanel('INFO')\">\n" +
@@ -382,12 +380,12 @@ angular.module("binding/zwave.tpl.html", []).run(["$templateCache", function($te
     "        </div>\n" +
     "\n" +
     "        <!-- Node Network Diagram -->\n" +
-    "        <div ng-show=\"deviceDisplay=='NETWORK'\" ng-style=\"styleList()\">\n" +
+    "        <div ng-show=\"deviceDisplay=='NETWORK'\" resize-panel>\n" +
     "            <vis-network data=\"networkNodes\" options=\"networkOptions\" events=\"networkEvents\"></vis-network>\n" +
     "        </div>\n" +
     "\n" +
     "        <!-- Footer -->\n" +
-    "        <div class=\"panel-footer panel-footer-small\">\n" +
+    "        <div class=\"panel-footer panel-footer-none\">\n" +
     "        </div>\n" +
     "    </div>\n" +
     "</div>\n" +
@@ -396,6 +394,7 @@ angular.module("binding/zwave.tpl.html", []).run(["$templateCache", function($te
 
 angular.module("dashboard/chart.tpl.html", []).run(["$templateCache", function($templateCache) {
   $templateCache.put("dashboard/chart.tpl.html",
+    "<!DOCTYPE html>\n" +
     "<!-- Chart and Item Lists -->\n" +
     "<div class=\"col-sm-4\">\n" +
     "    <div class=\"panel panel-default\">\n" +
@@ -485,8 +484,7 @@ angular.module("dashboard/chart.tpl.html", []).run(["$templateCache", function($
     "        </div>\n" +
     "\n" +
     "        <!-- Display the list of items -->\n" +
-    "        <div id=\"itemList\" class=\"list-group habmin-list\" ng-style=\"styleItemList()\" ng-if=\"!selectCharts\"\n" +
-    "             resize-page>\n" +
+    "        <div id=\"itemList\" class=\"list-group habmin-list\" ng-if=\"!selectCharts\" resize-panel>\n" +
     "            <div ng-if=\"itemsTotal>0\" ng-repeat=\"choice in items | filter:filterFunction\">\n" +
     "                <a role=\"presentation\" ng-hide=\"choice.filter\"\n" +
     "                   ng-class=\"{'list-group-item': true, 'active':choice.selected}\"\n" +
@@ -511,10 +509,9 @@ angular.module("dashboard/chart.tpl.html", []).run(["$templateCache", function($
     "        </div>\n" +
     "\n" +
     "        <!-- Display the list of predefined charts -->\n" +
-    "        <div id=\"chartList\" class=\"list-group habmin-list\" ng-style=\"styleChartList()\" ng-if=\"selectCharts\"\n" +
-    "             resize-page>\n" +
+    "        <div id=\"chartList\" class=\"list-group habmin-list\" ng-if=\"selectCharts\" resize-panel>\n" +
     "            <div ng-if=\"chartsTotal>0\" ng-repeat=\"choice in charts\">\n" +
-    "                <a role=\"presentation\" class=\"list-group-item\" ng-class=\"{'active':choice.selected=='yes'}\"\n" +
+    "                <a role=\"presentation\" ng-class=\"{'list-group-item': true, 'active': choice.selected=='yes'}\" ng-class=\"{'active':choice.selected=='yes'}\"\n" +
     "                   ng-click=\"selectChart(choice)\">\n" +
     "                    <habmin-icon class=\"icon-lg\" icon=\"{{choice.icon}}\"></habmin-icon>\n" +
     "                    <span>{{choice.name}}</span>\n" +
@@ -534,54 +531,40 @@ angular.module("dashboard/chart.tpl.html", []).run(["$templateCache", function($
     "                <span i18n=\"habmin.chartNoCharts\"></span>\n" +
     "            </div>\n" +
     "        </div>\n" +
-    "\n" +
-    "        <!-- Add a footer to the panel -->\n" +
-    "        <div class=\"panel-footer panel-footer-small\">\n" +
-    "            <div class=\"row\" ng-if=\"!selectCharts\">\n" +
-    "                <div class=\"col-xs-6\">\n" +
-    "                    <h6 class=\"text-center\">\n" +
-    "                        <span i18n=\"habmin.chartTotalItems\"></span>\n" +
-    "                        <span class=\"label label-info\">{{itemsTotal}}</span>\n" +
-    "                    </h6>\n" +
-    "                </div>\n" +
-    "                <div class=\"col-xs-6\">\n" +
-    "                    <h6 class=\"text-center\">\n" +
-    "                        <span i18n=\"habmin.chartSelectedItems\"></span>\n" +
-    "                        <span class=\"label label-success\">{{itemsSelected}}</span>\n" +
-    "                    </h6>\n" +
-    "                </div>\n" +
-    "            </div>\n" +
+    "        <!-- Footer -->\n" +
+    "        <div class=\"panel-footer panel-footer-none\">\n" +
     "        </div>\n" +
     "    </div>\n" +
     "</div>\n" +
     "\n" +
+    "<!-- Chart -->\n" +
     "<div class=\"col-sm-8\">\n" +
-    "    <div class=\"panel panel-default\" ng-style=\"styleChartPanel()\" resize-page>\n" +
-    "        <div class=\"panel-heading\" style=\"padding:3px;\" time-navigation>\n" +
-    "            <div class=\"btn-group btn-group-sm\">\n" +
-    "                <button type=\"button\" class=\"btn btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
+    "    <div class=\"panel panel-default\">\n" +
+    "        <div class=\"panel-heading\" style=\"padding:3px;\">\n" +
+    "            <div class=\"btn-group btn-group-sm\" role=\"group\">\n" +
+    "                <button type=\"button\" class=\"btn btn-sm btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
     "                        ng-click=\"stepWindow(-1)\">\n" +
     "                    <span class=\"fa fa-backward\"></span>\n" +
     "                </button>\n" +
-    "                <button type=\"button\" class=\"btn btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
+    "                <button type=\"button\" class=\"btn btn-sm btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
     "                        ng-click=\"stepWindow(1)\">\n" +
     "                    <span class=\"fa fa-forward\"></span>\n" +
     "                </button>\n" +
-    "                <button type=\"button\" class=\"btn btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
+    "                <button type=\"button\" class=\"btn btn-sm btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
     "                        ng-click=\"setNow()\">\n" +
     "                    <span class=\"fa fa-fast-forward\"></span>\n" +
     "                </button>\n" +
-    "                <button type=\"button\" class=\"btn btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
+    "                <button type=\"button\" class=\"btn btn-sm btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
     "                        ng-click=\"zoomWindow(0.10)\">\n" +
     "                    <span class=\"fa fa-search-minus\"></span>\n" +
     "                </button>\n" +
-    "                <button type=\"button\" class=\"btn btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
+    "                <button type=\"button\" class=\"btn btn-sm btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
     "                        ng-click=\"zoomWindow(-0.10)\">\n" +
     "                    <span class=\"fa fa-search-plus\"></span>\n" +
     "                </button>\n" +
     "            </div>\n" +
-    "            <div ng-class=\"{'btn-group btn-group-sm dropdown': true, 'open': calendarOpen}\">\n" +
-    "                <button type=\"button\" class=\"btn btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
+    "            <div ng-class=\"{'btn-group btn-group-sm': true, 'open': calendarOpen}\">\n" +
+    "                <button type=\"button\" class=\"btn btn-sm btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
     "                        ng-class=\"{active:graphWindow=='day'}\" ng-click=\"setWindow('day')\">\n" +
     "                        <span class=\"calendar-stack\">\n" +
     "                            <span class=\"fa fa-calendar-o\"></span>\n" +
@@ -589,7 +572,7 @@ angular.module("dashboard/chart.tpl.html", []).run(["$templateCache", function($
     "                        </span>\n" +
     "                    <span>Day</span>\n" +
     "                </button>\n" +
-    "                <button type=\"button\" class=\"btn btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
+    "                <button type=\"button\" class=\"btn btn-sm btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
     "                        ng-class=\"{active:graphWindow=='week'}\" ng-click=\"setWindow('week')\">\n" +
     "                        <span class=\"calendar-stack\">\n" +
     "                            <span class=\"fa fa-calendar-o\"></span>\n" +
@@ -597,7 +580,7 @@ angular.module("dashboard/chart.tpl.html", []).run(["$templateCache", function($
     "                        </span>\n" +
     "                    <span>Week</span>\n" +
     "                </button>\n" +
-    "                <button type=\"button\" class=\"btn btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
+    "                <button type=\"button\" class=\"btn btn-sm btn-default\" ng-disabled=\"!graphLoaded||calendarOpen\"\n" +
     "                        ng-class=\"{active:graphWindow=='month'}\" ng-click=\"setWindow('month')\">\n" +
     "                        <span class=\"calendar-stack\">\n" +
     "                            <span class=\"fa fa-calendar-o\"></span>\n" +
@@ -605,7 +588,7 @@ angular.module("dashboard/chart.tpl.html", []).run(["$templateCache", function($
     "                        </span>\n" +
     "                    <span>Month</span>\n" +
     "                </button>\n" +
-    "                <button type=\"button\" class=\"btn btn-default dropdown\" ng-disabled=\"!graphLoaded\"\n" +
+    "<!--                <button type=\"button\" class=\"btn btn-default dropdown\" ng-disabled=\"!graphLoaded\"\n" +
     "                        ng-init=\"calendarOpen=false\" ng-click=\"calendarOpen=!calendarOpen\">\n" +
     "                    <span class=\"fa fa-clock-o\"></span>&nbsp;Custom\n" +
     "                </button>\n" +
@@ -632,73 +615,7 @@ angular.module("dashboard/chart.tpl.html", []).run(["$templateCache", function($
     "                    <button class=\"btn btn-sm btn-primary\" ng-click=\"setDateRange();calendarOpen=false;\">OK</button>\n" +
     "                    <button class=\"btn btn-sm btn-warning\" ng-click=\"calendarOpen=false;\">Cancel</button>\n" +
     "                    </div>\n" +
-    "                </div>\n" +
-    "            </div>\n" +
-    "\n" +
-    "            <div class=\"btn-group btn-group-sm dropdown\">\n" +
-    "                <button type=\"button\" class=\"btn btn-default dropdown-toggle\" data-toggle=\"dropdown\"\n" +
-    "                        ng-disabled=\"!graphLoaded||calendarOpen\">\n" +
-    "                    <span class=\"fa fa-retweet\"></span>\n" +
-    "                    <span>Auto Update</span>\n" +
-    "                    <span class=\"caret\"></span>\n" +
-    "                </button>\n" +
-    "                <ul class=\"dropdown-menu\" role=\"menu\">\n" +
-    "                    <li>\n" +
-    "                        <a ng-click=\"setRefresh('0')\">\n" +
-    "                            <span ng-class=\"{'fa fa-fw': true, 'fa fa-check-square-o':refreshPeriod=='0'}\"></span>\n" +
-    "                            <span i18n=\"common.disabled\"></span>\n" +
-    "                        </a>\n" +
-    "                    </li>\n" +
-    "                    <li role=\"presentation\" class=\"divider\"></li>\n" +
-    "                    <li>\n" +
-    "                        <a ng-click=\"setRefresh('1.m')\">\n" +
-    "                            <span ng-class=\"{'fa fa-fw': true, 'fa fa-check-square-o':refreshPeriod=='1.m'}\"></span>\n" +
-    "                            <span i18n=\"habmin.period1Minute\"></span>\n" +
-    "                        </a>\n" +
-    "                    </li>\n" +
-    "                    <li>\n" +
-    "                        <a ng-click=\"setRefresh('2.m')\">\n" +
-    "                            <span ng-class=\"{'fa fa-fw': true, 'fa fa-check-square-o':refreshPeriod=='2.m'}\"></span>\n" +
-    "                            <span i18n=\"habmin.period2Minutes\"></span>\n" +
-    "                        </a>\n" +
-    "                    </li>\n" +
-    "                    <li>\n" +
-    "                        <a ng-click=\"setRefresh('5.m')\">\n" +
-    "                            <span ng-class=\"{'fa fa-fw': true, 'fa fa-check-square-o':refreshPeriod=='5.m'}\"></span>\n" +
-    "                            <span i18n=\"habmin.period5Minutes\"></span>\n" +
-    "                        </a>\n" +
-    "                    </li>\n" +
-    "                    <li>\n" +
-    "                        <a ng-click=\"setRefresh('10.m')\">\n" +
-    "                            <span ng-class=\"{'fa fa-fw': true, 'fa fa-check-square-o':refreshPeriod=='10.m'}\"></span>\n" +
-    "                            <span i18n=\"habmin.period10Minutes\"></span>\n" +
-    "                        </a>\n" +
-    "                    </li>\n" +
-    "                    <li>\n" +
-    "                        <a ng-click=\"setRefresh('15.m')\">\n" +
-    "                            <span ng-class=\"{'fa fa-fw': true, 'fa fa-check-square-o':refreshPeriod=='15.m'}\"></span>\n" +
-    "                            <span i18n=\"habmin.period15Minutes\"></span>\n" +
-    "                        </a>\n" +
-    "                    </li>\n" +
-    "                    <li>\n" +
-    "                        <a ng-click=\"setRefresh('20.m')\">\n" +
-    "                            <span ng-class=\"{'fa fa-fw': true, 'fa fa-check-square-o':refreshPeriod=='20.m'}\"></span>\n" +
-    "                            <span i18n=\"habmin.period20Minutes\"></span>\n" +
-    "                        </a>\n" +
-    "                    </li>\n" +
-    "                    <li>\n" +
-    "                        <a ng-click=\"setRefresh('30.m')\">\n" +
-    "                            <span ng-class=\"{'fa fa-fw': true, 'fa fa-check-square-o':refreshPeriod=='30.m'}\"></span>\n" +
-    "                            <span i18n=\"habmin.period30Minutes\"></span>\n" +
-    "                        </a>\n" +
-    "                    </li>\n" +
-    "                    <li>\n" +
-    "                        <a ng-click=\"setRefresh('1.h')\">\n" +
-    "                            <span ng-class=\"{'fa fa-fw': true, 'fa fa-check-square-o':refreshPeriod=='1.h'}\"></span>\n" +
-    "                            <span i18n=\"habmin.period1Hour\"></span>\n" +
-    "                        </a>\n" +
-    "                    </li>\n" +
-    "                </ul>\n" +
+    "                </div> -->\n" +
     "            </div>\n" +
     "\n" +
     "            <div ng-show=\"graphLoaded\" class=\"pull-right\">\n" +
@@ -706,7 +623,7 @@ angular.module("dashboard/chart.tpl.html", []).run(["$templateCache", function($
     "            </div>\n" +
     "        </div>\n" +
     "\n" +
-    "        <div class=\"panel-body\" ng-style=\"styleChart()\" style=\"padding:3px;\">\n" +
+    "        <div class=\"panel-body\" style=\"padding:3px;\" resize-panel>\n" +
     "            <vis-graph2d ng-show=\"graphLoaded\" data=\"graphData\" options=\"graphOptions\"\n" +
     "                         events=\"graphEvents\"></vis-graph2d>\n" +
     "        </div>\n" +
@@ -717,6 +634,7 @@ angular.module("dashboard/chart.tpl.html", []).run(["$templateCache", function($
 
 angular.module("dashboard/chartSave.tpl.html", []).run(["$templateCache", function($templateCache) {
   $templateCache.put("dashboard/chartSave.tpl.html",
+    "<!DOCTYPE html>\n" +
     "<div class=\"modal-header\">\n" +
     "    <h3 class=\"modal-title\" i18n=\"habmin.chartSaveChartTitle\"></h3>\n" +
     "</div>\n" +
@@ -757,8 +675,8 @@ angular.module("dashboard/chartSave.tpl.html", []).run(["$templateCache", functi
     "    </div>\n" +
     "</div>\n" +
     "<div class=\"modal-footer\">\n" +
-    "    <button class=\"btn btn-primary\" ng-click=\"ok()\" i18n=\"common.save\"></button>\n" +
-    "    <button class=\"btn btn-warning\" ng-click=\"cancel()\" i18n=\"common.cancel\"></button>\n" +
+    "    <button type=\"button\" class=\"btn btn-primary\" ng-click=\"ok()\" i18n=\"common.save\"></button>\n" +
+    "    <button type=\"button\" class=\"btn btn-warning\" ng-click=\"cancel()\" i18n=\"common.cancel\"></button>\n" +
     "</div>\n" +
     "");
 }]);
@@ -838,7 +756,7 @@ angular.module("dashboard/chartSaveGeneral.tpl.html", []).run(["$templateCache",
     "    <div class=\"form-group\">\n" +
     "        <label for=\"inputPeriod\" class=\"control-label\" i18n=\"habmin.chartSavePeriod\"></label>\n" +
     "        <div>\n" +
-    "            <select selectpicker ng-model=\"model.period\" class=\"form-control\" id=\"inputPeriod\">\n" +
+    "            <select ng-model=\"model.period\" class=\"form-control\" id=\"inputPeriod\">\n" +
     "                <option value=\"3600\" i18n=\"habmin.period1Hour\"></option>\n" +
     "                <option value=\"7200\" i18n=\"habmin.period2Hours\"></option>\n" +
     "                <option value=\"10800\" i18n=\"habmin.period3Hours\"></option>\n" +
@@ -1003,6 +921,14 @@ angular.module("home/home.tpl.html", []).run(["$templateCache", function($templa
     "        A graphical user interface for the OpenHAB Home Automation System.\n" +
     "    </p>\n" +
     "\n" +
+    "    <p class=\"text-center\">\n" +
+    "        <small>\n" +
+    "            <p>Version: {{version}}, {{date}}</p>\n" +
+    "            <p>Phone: {{phone}}</p>\n" +
+    "            <p>Server: {{server}}</p>\n" +
+    "        </small>\n" +
+    "    </p>\n" +
+    "\n" +
     "    <div class=\"text-center\">\n" +
     "        <div class=\"btn-group\">\n" +
     "            <a href=\"https://github.com/cdjackson/HABmin/wiki\" class=\"btn btn-large btn-default\">\n" +
@@ -1049,7 +975,7 @@ angular.module("home/home.tpl.html", []).run(["$templateCache", function($templa
 
 angular.module("sitemap/sitemap.tpl.html", []).run(["$templateCache", function($templateCache) {
   $templateCache.put("sitemap/sitemap.tpl.html",
-    "<div><div dynamic-sitemap></div></div>\n" +
+    "<div dynamic-sitemap></div>\n" +
     "");
 }]);
 
@@ -1065,10 +991,7 @@ angular.module("user/userChart.tpl.html", []).run(["$templateCache", function($t
     "            <label for=\"inputDefaultPeriod\" class=\"col-sm-3 control-label\" i18n=\"habmin.userChartPeriod\"></label>\n" +
     "\n" +
     "            <div class=\"col-sm-9\">\n" +
-    "                <select class=\"form-control\" id=\"inputDefaultPeriod\"\n" +
-    "                        ng-model=\"selected\"\n" +
-    "                        ng-option=\"c.value as c.label for c in periodOptions\"\n" +
-    "                        selectpicker=\"xx\" toggle-dropdown live-search=\"false\">\n" +
+    "                <select class=\"form-control\" id=\"inputDefaultPeriod\">\n" +
     "                    <option value=\"-1\">hello</option>\n" +
     "                    <option value=\"3600\" i18n=\"habmin.period1Hour\"></option>\n" +
     "                    <option value=\"1\">hello 2</option>\n" +
@@ -1082,15 +1005,13 @@ angular.module("user/userChart.tpl.html", []).run(["$templateCache", function($t
     "            <div class=\"col-sm-9\">\n" +
     "                <input type=\"checkbox\">\n" +
     "            </div>\n" +
-    "            {{test}}\n" +
-    "            {{periodOptions}}\n" +
     "        </div>\n" +
     "    </form>\n" +
     "\n" +
     "</div>\n" +
     "<div class=\"modal-footer\">\n" +
-    "    <button class=\"btn btn-primary\" ng-click=\"ok()\" i18n=\"common.save\"></button>\n" +
-    "    <button class=\"btn btn-warning\" ng-click=\"cancel()\" i18n=\"common.cancel\"></button>\n" +
+    "    <button type=\"button\" class=\"btn btn-primary\" ng-click=\"ok()\" i18n=\"common.save\"></button>\n" +
+    "    <button type=\"button\" class=\"btn btn-warning\" ng-click=\"cancel()\" i18n=\"common.cancel\"></button>\n" +
     "</div>\n" +
     "");
 }]);
@@ -1102,14 +1023,18 @@ angular.module("user/userGeneral.tpl.html", []).run(["$templateCache", function(
     "</div>\n" +
     "<div class=\"modal-body\">\n" +
     "    <form class=\"form-horizontal\" role=\"form\">\n" +
-    "{{embedded}}\n" +
-    "        <div ng-show=\"embedded\" class=\"form-group\">\n" +
-    "            <label for=\"inputEmail2\" class=\"col-sm-3 control-label\">Name</label>\n" +
+    "        <div class=\"form-group\">\n" +
+    "            <label for=\"inputTheme\" class=\"col-sm-3 control-label\">Theme</label>\n" +
     "\n" +
     "            <div class=\"col-sm-9\">\n" +
-    "                <input type=\"email\" class=\"form-control\" id=\"inputEmail2\" placeholder=\"Name\">\n" +
+    "                <select class=\"form-control\" id=\"inputTheme\" ng-model=\"model.theme\">\n" +
+    "                    <option value=\"slate\">Slate</option>\n" +
+    "                    <option value=\"paper\">Paper</option>\n" +
+    "                    <option value=\"yeti\">Yeti</option>\n" +
+    "                </select>\n" +
     "            </div>\n" +
     "        </div>\n" +
+    "\n" +
     "        <div class=\"form-group\">\n" +
     "            <label for=\"inputEmail2\" class=\"col-sm-3 control-label\">Name</label>\n" +
     "\n" +
@@ -1138,22 +1063,12 @@ angular.module("user/userGeneral.tpl.html", []).run(["$templateCache", function(
     "                <input type=\"password\" class=\"form-control\" id=\"inputMobile\" placeholder=\"Phone\">\n" +
     "            </div>\n" +
     "        </div>\n" +
-    "\n" +
-    "        <div class=\"form-group\">\n" +
-    "            <div class=\"col-sm-offset-3 col-sm-10\">\n" +
-    "                <div class=\"checkbox\">\n" +
-    "                    <label>\n" +
-    "                        <input type=\"checkbox\"> Remember me\n" +
-    "                    </label>\n" +
-    "                </div>\n" +
-    "            </div>\n" +
-    "        </div>\n" +
     "    </form>\n" +
     "\n" +
     "</div>\n" +
     "<div class=\"modal-footer\">\n" +
-    "    <button class=\"btn btn-primary\" ng-click=\"ok()\" i18n=\"common.save\"></button>\n" +
-    "    <button class=\"btn btn-warning\" ng-click=\"cancel()\" i18n=\"common.cancel\"></button>\n" +
+    "    <button type=\"button\" class=\"btn btn-primary\" ng-click=\"ok()\" i18n=\"common.save\"></button>\n" +
+    "    <button type=\"button\" class=\"btn btn-warning\" ng-click=\"cancel()\" i18n=\"common.cancel\"></button>\n" +
     "</div>\n" +
     "");
 }]);