+++ CHANGELOG.md
@@ -1,15 +1,38 @@
-# 0.0.2 (2014-12-26)
+# 0.0.2 (2014-12-28)
 
 ## Features
 ### Mobile
 
 *  Initial addition of native mobile versions of HABmin ([d9a403ac](git@github.com:cdjackson/HABmin2/commit/d9a403ac))
 
+* Persist server address to local storage ([8a619066](git@github.com:cdjackson/HABmin2/commit/8a619066))
+
 ### UI
 
 * Add theme switching ([eebb64d3](git@github.com:cdjackson/HABmin2/commit/eebb64d3))
 
+* Persist user theme choice to local storage and reset on startup ([2ce3952b](git@github.com:cdjackson/HABmin2/commit/2ce3952b))
 
+* Add weather-icons ([ae1ce462](git@github.com:cdjackson/HABmin2/commit/ae1ce462))
+
+* Improve styling of sitemap widgets ([de1896cb](git@github.com:cdjackson/HABmin2/commit/de1896cb))
+
+
+
+## Bug fixes
+### ZWave
+
+* Fix network diagram resizing ([b875d206](git@github.com:cdjackson/HABmin2/commit/b875d206))
+
+* Fix battery icon colour when battery level <40% ([1aa5fdd9](git@github.com:cdjackson/HABmin2/commit/1aa5fdd9))
+
+* Fix device status if retries is greater than 100% ([5a48ff92](git@github.com:cdjackson/HABmin2/commit/5a48ff92))
+
+### UI
+
+* Update chart series icon styles ([64485dc5](git@github.com:cdjackson/HABmin2/commit/64485dc5))
+
+* Fix error with login and authorisation window ([9a0be091](git@github.com:cdjackson/HABmin2/commit/9a0be091))